requirement_loader.txt
anyio==3.7.0
async-timeout==4.0.2
exceptiongroup==1.1.1
idna==3.4
redis==4.5.5
sniffio==1.3.0
starlette==0.28.0
websocket-client==1.5.3
